+++ src/commands/buildReleases.ts
@@ -1,7 +1,6 @@
 import chalk from "chalk";
 import { execCommand } from "../utils/execHelpers.js";
 import { LoggerHelpers } from "../utils/loggerHelpers.js";
-import { incrementIosBuildVersion } from "../utils/iosBuildUpdate.js";
 
 export {
   buildFlutterApk,
@@ -57,8 +56,6 @@ async function buildFlutterBundle(noFvm: boolean) {
 }
 
 async function buildFlutterIos(noFvm: boolean) {
-  await incrementIosBuildVersion();
-
   LoggerHelpers.info(
     noFvm
       ? "Building Flutter iOS app without FVM..."
@@ -82,8 +79,6 @@ async function buildFlutterIos(noFvm: boolean) {
 }
 
 async function buildFlutterIpa(noFvm: boolean) {
-  await incrementIosBuildVersion();
-
   LoggerHelpers.info(
     noFvm
       ? "Creating release IPA without FVM..."