+++ CHANGELOG.md
@@ -13,9 +13,9 @@ Each section lists the changes in chronological order, with the most recent rele
 ## [1.1.0] - 2024-11-21
 ### Added
 - New Commands:
+  - `sheno flutter-update-version`: Update the version and build numbers for both Android and iOS, with support for automatic versioning and build number management.
   - `sheno flutter-build-ios`: Build the Flutter iOS app for release, with support for FVM and automatic version incrementing
   - `sheno flutter-build-ipa`: Create a release IPA for the Flutter app, with support for FVM and automatic version incrementing.
-  - Enhanced iOS build workflows to include automatic version number incrementation before building.
 
 ## [1.0.0] - 2024-11-18
 ### Added